+++ constructs/revali_client/revali_client/lib/revali_client.dart
@@ -14,3 +14,4 @@ export 'src/revali_client.dart';
 export 'src/server_exception.dart';
 export 'src/session_storage.dart';
 export 'src/storage.dart';
+export 'src/stream_errors.dart';